Move services to core/services
Core folder is for singleton services possibly accessed from anywhere in
the App and non-lazy-loading modules
